--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,4 @@ module github.com/dwisiswant0/pcregexp
 
 go 1.18
 
-require github.com/ebitengine/purego v0.8.2
+require github.com/ebitengine/purego v0.8.3