--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/dwisiswant0/pcregexp
 
-go 1.18
+go 1.20
 
 require github.com/ebitengine/purego v0.8.3