changed:

* adds JIT options & compilation supports
* BREAKING: replace obsolete `pcre2_set_recursion_limit` with `pcre2_set_depth_limit` (`RecursionLimit` => `DepthLimit`)
* cache matches data
* exports `NeedsPCRE` util