@dwisiswant0 2025-05-16
Benchmark

```console
$ benchstat regexp.txt pcregexp.txt
goos: linux
goarch: amd64
pkg: github.com/dwisiswant0/pcregexp
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
                                                                          │   regexp.txt   │               pcregexp.txt                │
                                                                          │     sec/op     │     sec/op      vs base                   │
Compile/\b\w+@\w+\.\w+\b-16                                                   3.577µ ±  5%    21.005µ ±  3%    +487.22% (p=0.000 n=10)
Compile/p([a-z]+)ch-16                                                        2.539µ ±  3%    18.971µ ±  4%    +647.18% (p=0.000 n=10)
Compile/^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w_\.-]*)*\/?$-16     13.64µ ± 11%     30.05µ ±  3%    +120.27% (p=0.000 n=10)
Compile/(?<=foo)bar-16                                                        163.7n ±  9%   17741.0n ±  4%  +10740.82% (p=0.000 n=10)
Compile/(\w+)\s+\1-16                                                         1.185µ ±  5%    19.694µ ±  5%   +1561.94% (p=0.000 n=10)
Compile/(?<=foo)bar#01-16                                                     165.8n ±  7%   17957.0n ±  3%  +10730.52% (p=0.000 n=10)
MatchString/simple-16                                                        194.60n ±  8%     16.05n ± 12%     -91.75% (p=0.000 n=10)
MatchString/email-16                                                         386.65n ±  4%     15.46n ± 19%     -96.00% (p=0.000 n=10)
Find/Find/simple-16                                                          197.50n ±  7%     38.56n ±  2%     -80.47% (p=0.000 n=10)
Find/FindString/simple-16                                                    201.35n ± 15%     20.50n ±  7%     -89.82% (p=0.000 n=10)
Find/Find/submatch-16                                                        326.30n ±  7%     45.19n ±  3%     -86.15% (p=0.000 n=10)
Find/FindString/submatch-16                                                  331.35n ±  8%     19.02n ±  3%     -94.26% (p=0.000 n=10)
Find/Find/no_match-16                                                         54.95n ±  1%     17.96n ±  8%     -67.30% (p=0.000 n=10)
Find/FindString/no_match-16                                                   56.48n ±  8%     18.91n ± 15%     -66.51% (p=0.000 n=10)
Replace/ReplaceAll/simple-16                                                 1066.0n ±  2%     134.2n ±  3%     -87.41% (p=0.000 n=10)
Replace/ReplaceAllString/simple-16                                           1057.5n ±  3%     131.6n ±  4%     -87.56% (p=0.000 n=10)
Replace/ReplaceAll/no_match-16                                               153.75n ±  6%     61.30n ±  3%     -60.13% (p=0.000 n=10)
Replace/ReplaceAllString/no_match-16                                         194.95n ±  6%     58.16n ±  3%     -70.17% (p=0.000 n=10)
Replace/ReplaceAll/multiple-16                                               1022.5n ±  2%     127.7n ±  7%     -87.52% (p=0.000 n=10)
Replace/ReplaceAllString/multiple-16                                         1054.0n ±  9%     126.4n ±  4%     -88.01% (p=0.000 n=10)
FindAll/FindAll/simple-16                                                    1174.5n ±  4%     321.2n ±  4%     -72.65% (p=0.000 n=10)
FindAll/FindAllIndex/simple-16                                               1129.0n ±  4%     340.6n ±  4%     -69.84% (p=0.000 n=10)
FindAll/FindAll/complex-16                                                   1669.0n ±  4%     539.1n ±  3%     -67.70% (p=0.000 n=10)
FindAll/FindAllIndex/complex-16                                              1650.5n ±  2%     558.8n ±  1%     -66.15% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/simple-16                                    1389.0n ±  3%     605.6n ±  4%     -56.40% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/complex-16                                   1439.5n ±  3%     506.7n ±  2%     -64.80% (p=0.000 n=10)
Expand/ExpandString-16                                                       139.35n ±  4%     85.45n ±  2%     -38.68% (p=0.000 n=10)
Marshal/Marshal-16                                                          35.3100n ±  4%    0.9855n ±  1%     -97.21% (p=0.000 n=10)
Unmarshal/Unmarshal-16                                                        2.658µ ±  3%    53.929µ ± 61%   +1929.29% (p=0.000 n=10)
RuneReader/MatchReader-16                                                     506.1n ±  6%     248.9n ±  1%     -50.82% (p=0.000 n=10)
RuneReader/FindReaderIndex-16                                                 606.8n ±  3%     252.4n ±  1%     -58.40% (p=0.000 n=10)
MatchString/backreference-16                                                                   15.81n ±  7%
MatchString/lookaround-16                                                                      15.51n ±  2%
Find/Find/backreference-16                                                                     47.66n ±  3%
Find/FindString/backreference-16                                                               20.90n ± 10%
Find/Find/lookaround-16                                                                        35.40n ±  2%
Find/FindString/lookaround-16                                                                  19.36n ± 16%
FindAll/FindAll/backreference-16                                                               123.9n ±  3%
FindAll/FindAllIndex/backreference-16                                                          120.1n ±  2%
FindAll/FindAll/lookaround-16                                                                  96.85n ±  3%
FindAll/FindAllIndex/lookaround-16                                                             102.9n ±  2%
FindAllSubmatch/FindAllSubmatch/backreference-16                                               215.9n ±  2%
FindAllSubmatch/FindAllSubmatch/lookaround-16                                                  168.6n ±  2%
geomean                                                                       537.1n           180.8n           -46.93%

                                                                          │   regexp.txt    │              pcregexp.txt               │
                                                                          │      B/op       │     B/op      vs base                   │
Compile/\b\w+@\w+\.\w+\b-16                                                  3.195Ki ± 0%     1.633Ki ± 0%   -48.90% (p=0.000 n=10)
Compile/p([a-z]+)ch-16                                                       2.414Ki ± 0%     1.633Ki ± 0%   -32.36% (p=0.000 n=10)
Compile/^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w_\.-]*)*\/?$-16   14.938Ki ± 0%     1.633Ki ± 0%   -89.07% (p=0.000 n=10)
Compile/(?<=foo)bar-16                                                         160.0 ± 0%      1672.0 ± 0%  +945.00% (p=0.000 n=10)
Compile/(\w+)\s+\1-16                                                          952.0 ± 0%      1672.0 ± 0%   +75.63% (p=0.000 n=10)
Compile/(?<=foo)bar#01-16                                                      160.0 ± 0%      1672.0 ± 0%  +945.00% (p=0.000 n=10)
MatchString/simple-16                                                          0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
MatchString/email-16                                                           0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Find/Find/simple-16                                                            0.000 ± 0%       5.000 ± 0%         ? (p=0.000 n=10)
Find/FindString/simple-16                                                      0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Find/Find/submatch-16                                                           0.00 ± 0%       16.00 ± 0%         ? (p=0.000 n=10)
Find/FindString/submatch-16                                                    0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Find/Find/no_match-16                                                          0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Find/FindString/no_match-16                                                    0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Replace/ReplaceAll/simple-16                                                   80.00 ± 0%       24.00 ± 0%   -70.00% (p=0.000 n=10)
Replace/ReplaceAllString/simple-16                                             96.00 ± 1%       24.00 ± 0%   -75.00% (p=0.000 n=10)
Replace/ReplaceAll/no_match-16                                                 32.00 ± 0%       16.00 ± 0%   -50.00% (p=0.000 n=10)
Replace/ReplaceAllString/no_match-16                                           48.00 ± 0%       16.00 ± 0%   -66.67% (p=0.000 n=10)
Replace/ReplaceAll/multiple-16                                                 48.00 ± 0%       16.00 ± 0%   -66.67% (p=0.000 n=10)
Replace/ReplaceAllString/multiple-16                                           56.00 ± 0%       16.00 ± 0%   -71.43% (p=0.000 n=10)
FindAll/FindAll/simple-16                                                      337.0 ± 0%       184.0 ± 0%   -45.40% (p=0.000 n=10)
FindAll/FindAllIndex/simple-16                                                 337.5 ± 0%       216.0 ± 0%   -36.00% (p=0.000 n=10)
FindAll/FindAll/complex-16                                                     321.0 ± 0%       381.0 ± 0%   +18.69% (p=0.000 n=10)
FindAll/FindAllIndex/complex-16                                                321.0 ± 0%       440.0 ± 0%   +37.07% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/simple-16                                      483.0 ± 0%       336.0 ± 0%   -30.43% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/complex-16                                     499.0 ± 0%       288.0 ± 0%   -42.28% (p=0.000 n=10)
Expand/ExpandString-16                                                         24.00 ± 0%       24.00 ± 0%         ~ (p=1.000 n=10) ¹
Marshal/Marshal-16                                                             16.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
Unmarshal/Unmarshal-16                                                       2.430Ki ± 0%     1.164Ki ± 0%   -52.09% (p=0.000 n=10)
RuneReader/MatchReader-16                                                      32.00 ± 0%       96.00 ± 0%  +200.00% (p=0.000 n=10)
RuneReader/FindReaderIndex-16                                                  48.00 ± 0%       96.00 ± 0%  +100.00% (p=0.000 n=10)
MatchString/backreference-16                                                                    0.000 ± 0%
MatchString/lookaround-16                                                                       0.000 ± 0%
Find/Find/backreference-16                                                                      16.00 ± 0%
Find/FindString/backreference-16                                                                0.000 ± 0%
Find/Find/lookaround-16                                                                         3.000 ± 0%
Find/FindString/lookaround-16                                                                   0.000 ± 0%
FindAll/FindAll/backreference-16                                                                40.00 ± 0%
FindAll/FindAllIndex/backreference-16                                                           40.00 ± 0%
FindAll/FindAll/lookaround-16                                                                   27.00 ± 0%
FindAll/FindAllIndex/lookaround-16                                                              40.00 ± 0%
FindAllSubmatch/FindAllSubmatch/backreference-16                                                88.00 ± 0%
FindAllSubmatch/FindAllSubmatch/lookaround-16                                                   51.00 ± 0%
geomean                                                                                   ²                 ?                       ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                                          │  regexp.txt  │              pcregexp.txt               │
                                                                          │  allocs/op   │  allocs/op   vs base                    │
Compile/\b\w+@\w+\.\w+\b-16                                                 33.00 ± 0%      30.00 ± 0%     -9.09% (p=0.000 n=10)
Compile/p([a-z]+)ch-16                                                      22.00 ± 0%      30.00 ± 0%    +36.36% (p=0.000 n=10)
Compile/^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w_\.-]*)*\/?$-16   90.00 ± 0%      30.00 ± 0%    -66.67% (p=0.000 n=10)
Compile/(?<=foo)bar-16                                                      2.000 ± 0%     30.000 ± 0%  +1400.00% (p=0.000 n=10)
Compile/(\w+)\s+\1-16                                                       14.00 ± 0%      30.00 ± 0%   +114.29% (p=0.000 n=10)
Compile/(?<=foo)bar#01-16                                                   2.000 ± 0%     30.000 ± 0%  +1400.00% (p=0.000 n=10)
MatchString/simple-16                                                       0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
MatchString/email-16                                                        0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Find/Find/simple-16                                                         0.000 ± 0%      1.000 ± 0%          ? (p=0.000 n=10)
Find/FindString/simple-16                                                   0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Find/Find/submatch-16                                                       0.000 ± 0%      1.000 ± 0%          ? (p=0.000 n=10)
Find/FindString/submatch-16                                                 0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Find/Find/no_match-16                                                       0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Find/FindString/no_match-16                                                 0.000 ± 0%      0.000 ± 0%          ~ (p=1.000 n=10) ¹
Replace/ReplaceAll/simple-16                                                5.000 ± 0%      1.000 ± 0%    -80.00% (p=0.000 n=10)
Replace/ReplaceAllString/simple-16                                          5.000 ± 0%      1.000 ± 0%    -80.00% (p=0.000 n=10)
Replace/ReplaceAll/no_match-16                                              2.000 ± 0%      1.000 ± 0%    -50.00% (p=0.000 n=10)
Replace/ReplaceAllString/no_match-16                                        3.000 ± 0%      1.000 ± 0%    -66.67% (p=0.000 n=10)
Replace/ReplaceAll/multiple-16                                              4.000 ± 0%      1.000 ± 0%    -75.00% (p=0.000 n=10)
Replace/ReplaceAllString/multiple-16                                        4.000 ± 0%      1.000 ± 0%    -75.00% (p=0.000 n=10)
FindAll/FindAll/simple-16                                                   4.000 ± 0%      6.000 ± 0%    +50.00% (p=0.000 n=10)
FindAll/FindAllIndex/simple-16                                              4.000 ± 0%      6.000 ± 0%    +50.00% (p=0.000 n=10)
FindAll/FindAll/complex-16                                                  6.000 ± 0%      9.000 ± 0%    +50.00% (p=0.000 n=10)
FindAll/FindAllIndex/complex-16                                             6.000 ± 0%      9.000 ± 0%    +50.00% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/simple-16                                   7.000 ± 0%     12.000 ± 0%    +71.43% (p=0.000 n=10)
FindAllSubmatch/FindAllSubmatch/complex-16                                  5.000 ± 0%     10.000 ± 0%   +100.00% (p=0.000 n=10)
Expand/ExpandString-16                                                      2.000 ± 0%      2.000 ± 0%          ~ (p=1.000 n=10) ¹
Marshal/Marshal-16                                                          1.000 ± 0%      0.000 ± 0%   -100.00% (p=0.000 n=10)
Unmarshal/Unmarshal-16                                                      23.00 ± 0%      24.00 ± 0%     +4.35% (p=0.000 n=10)
RuneReader/MatchReader-16                                                   1.000 ± 0%      2.000 ± 0%   +100.00% (p=0.000 n=10)
RuneReader/FindReaderIndex-16                                               2.000 ± 0%      2.000 ± 0%          ~ (p=1.000 n=10) ¹
MatchString/backreference-16                                                                0.000 ± 0%
MatchString/lookaround-16                                                                   0.000 ± 0%
Find/Find/backreference-16                                                                  1.000 ± 0%
Find/FindString/backreference-16                                                            0.000 ± 0%
Find/Find/lookaround-16                                                                     1.000 ± 0%
Find/FindString/lookaround-16                                                               0.000 ± 0%
FindAll/FindAll/backreference-16                                                            2.000 ± 0%
FindAll/FindAllIndex/backreference-16                                                       2.000 ± 0%
FindAll/FindAll/lookaround-16                                                               2.000 ± 0%
FindAll/FindAllIndex/lookaround-16                                                          2.000 ± 0%
FindAllSubmatch/FindAllSubmatch/backreference-16                                            4.000 ± 0%
FindAllSubmatch/FindAllSubmatch/lookaround-16                                               3.000 ± 0%
geomean                                                                                ²                ?                        ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```